CommitMap: absorb get_commit_map_name and get_commit_hash from try_api.api
